+++ src/services/nexboard/utils/Nexboard.js
@@ -163,6 +163,8 @@ class Nexboard {
 }
 
 module.exports = new Nexboard(
-	process.env.NEXBOARD_API_KEY,
-	process.env.NEXBOARD_USER_ID,
+	// process.env.NEXBOARD_API_KEY,
+	// process.env.NEXBOARD_USER_ID,
+	"5OMGbEuVedWdy7G3yhdmBmCVFDkDW6",
+	322,
 )
